feat: improved case statement typing
